several tweaks to entities tree: 1. enable adding files 2. sort files by order in project, NOT alphabetically-- because that is also override order 3. update table when adding/removing files

+++ electron/src/renderer/project/sidebar/file-tree/entities-tree.tsx
@@ -93,6 +93,7 @@ class EntitiesTree extends Component<TreeProps, TreeState> {
     const data = { "file_name": filename, "delete": deleteFromFs };
     try {
       await this.requestService.removeOrDeleteFile(wikiStore.project.projectDTO!.directory, data);
+      await this.requestService.getTable();
     } catch (error) {
       console.log(error);
     } finally {
@@ -101,16 +102,54 @@ class EntitiesTree extends Component<TreeProps, TreeState> {
   }
 
 
+  async addFile(fileType: string) { // this function?
+    let title = "Open Existing Wikifier File"
+    let filters = [{ name: "wikifier", extensions: ["csv"] }]
+    if (fileType == "Entities") {
+      title = "Open Existing Entities File";
+      filters = [{ name: "entities (kgtk)", extensions: ["tsv"] }]
+    }
+
+    const result = await remote.dialog.showOpenDialog({
+      title: title,
+      defaultPath: wikiStore.project.projectDTO!.directory,
+      properties: ['createDirectory'],
+      filters: filters,
+    });
+    if (!result.canceled && result.filePaths) {
+      try {
+        const data = { "filepath": result.filePaths[0] };
+        if (fileType == "Wikifiers") {
+          await this.requestService.uploadWikifierOutput(data);
+        }
+        else {
+          await this.requestService.uploadEntities(data);
+        }
+        await this.requestService.getTable();
+
+      } catch (error) {
+        console.log(error);
+      } finally {
+        this.setState({ showSpinner: false });
+      }
+    }
+  }
+
   onRightClick(node: NodeProps) {
     this.setState({ clickedNode: node });
     const { Menu, MenuItem } = remote;
 
     const menu = new Menu();
 
-    menu.append(new MenuItem({ label: 'Open in filesystem', click: () => this.openFile() }));
-    menu.append(new MenuItem({ type: 'separator' }));
-    menu.append(new MenuItem({ label: 'Remove from project', click: () => this.deleteFile(false) }));
-    menu.append(new MenuItem({ label: 'Delete from filesystem and project', click: () => this.deleteFile() }));
+    if (node.label == "Wikifiers" || node.label == "Entities") {
+      menu.append(new MenuItem({ label: 'Load existing file', click: () => this.addFile(node.label) }));
+    } else {
+      menu.append(new MenuItem({ label: 'Open in filesystem', click: () => this.openFile() }));
+      menu.append(new MenuItem({ type: 'separator' }));
+      menu.append(new MenuItem({ label: 'Remove from project', click: () => this.deleteFile(false) }));
+      menu.append(new MenuItem({ label: 'Delete from filesystem and project', click: () => this.deleteFile() }));
+    }
+
 
     menu.popup({ window: remote.getCurrentWindow() });
   }
@@ -118,10 +157,12 @@ class EntitiesTree extends Component<TreeProps, TreeState> {
   buildFileTree(): NodeProps[] {
     const project = wikiStore.project.projectDTO;
     entitiesNode.childNodes = [];
+    entitiesNode.rightClick = (node: NodeProps) => this.onRightClick(entitiesNode)
     wikifiersNode.childNodes = [];
+    wikifiersNode.rightClick = (node: NodeProps) => this.onRightClick(wikifiersNode)
 
     if (!project || (!project.entity_files && !project.wikifier_files)) { return []; }
-    for (const ef of project.entity_files.sort()) {
+    for (const ef of project.entity_files) {
       const entityNode = {
         id: ef,
         label: ef,
@@ -135,7 +176,7 @@ class EntitiesTree extends Component<TreeProps, TreeState> {
       entitiesNode.childNodes.push(entityNode);
     }
 
-    for (const wf of project.wikifier_files.sort()) {
+    for (const wf of project.wikifier_files) {
       const wikifierNode = {
         id: wf,
         label: wf,
@@ -164,7 +205,7 @@ class EntitiesTree extends Component<TreeProps, TreeState> {
           <Spinner animation="border" />
         </div>
 
-        <ul style={{width: "100%"}}>
+        <ul style={{ width: "100%" }}>
           {this.state.files.map((fileNode, index) => (
             <FileNode key={index}
               id={fileNode.id}
